# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: API Rest - Backend - Arena Mafia
  version: v1
  description: |
    Backend do projeto para a Arena Mafia, desenvolvido como parte do curso de
    Análise e Desenvolvimento de Sistemas do Instituto Federal de Santa Catarina.
  termsOfService: https://github.com/LucasRibasCardoso/backend-arena-mafia
  license:
    name: Apache 2.0
    url: https://github.com/LucasRibasCardoso
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token de autenticação JWT obtido no login."
  schemas:
    FieldErrorResponseDto:
      type: object
      properties:
        fieldName: {
          type: string,
          description: "Nome do campo que falhou na validação.",
          example: "username"
        }
        message: {
          type: string,
          description: "Mensagem de erro para o campo.",
          example: "O usuário é obrigatório"
        }
      required: [ fieldName, message ]
    ErrorResponseDto:
      type: object
      properties:
        timestamp: {
          type: string,
          format: date-time,
          example: "2025-08-15T13:55:00Z"
        }
        status: {
          type: integer,
          example: 400
        }
        message: {
          type: string,
          example: "Requisição inválida."
        }
        path: {
          type: string,
          example: "/api/some/endpoint"
        }
        fieldErrors:
          type: array
          description: "Lista de erros de validação de campos, se houver."
          items: { $ref: "#/components/schemas/FieldErrorResponseDto" }
          nullable: true
      required: [ timestamp, status, message, path, fieldErrors ]
    LoginRequestDto:
      type: object
      required: [ username, password ]
      properties:
        username: {
          type: string,
          description: "Nome de usuário.",
          example: "testuser"
        }
        password: {
          type: string,
          description: "Senha do usuário.",
          example: "123456"
        }
    TokenResponseDto:
      type: object
      properties:
        phone: {
          type: string,
          description: "Telefone do usuário autenticado.",
          example: "+5547912345678"
        }
        username: {
          type: string,
          description: "Nome de usuário autenticado.",
          example: "usuario123"
        }
        fullName: {
          type: string,
          description: "Nome completo do usuário autenticado.",
          example: "João da Silva"
        }
        role: {
          type: string,
          description: "Permissão do usuário autenticado.",
          example: "ROLE_USER"
        }
        accessToken: {
          type: string,
          description: "JWT de acesso.",
          example: "eyJhbGci..."
        }
    SignupRequestDto:
      type: object
      description: "Dados necessários para o cadastro de um novo usuário. A senha e a confirmação de senha devem ser idênticas."
      required: [ username, fullName, phone, password, confirmPassword ]
      properties:
        username:
          type: string
          description: "Nome de usuário único, deve conter apenas letras, números e underscore (_)."
          example: "joao_silva99"
        fullName:
          type: string
          description: "Nome completo do usuário."
          example: "João da Silva"
        phone:
          type: string
          description: "Número de telefone no formato internacional, incluindo o código do país e DDD."
          example: "+5547912345678"
        password:
          type: string
          format: "password"
          description: "Senha do usuário. Deve ter entre 6 e 20 caracteres."
          example: "senhaForte123"
        confirmPassword:
          type: string
          format: "password"
          description: "Confirmação da senha. Deve ser idêntica ao campo 'password'."
          example: "senhaForte123"
    SignupResponseDto:
      type: object
      properties:
        status:
          type: string
          description: "Status da conta do usuário após o registro."
          example: "pending_verification"
        message:
          type: string
          description: "Mensagem de confirmação do sucesso da operação."
          example: "Conta criada com sucesso. Um código de verificação foi enviado para o seu telefone."
        identifier:
          type: string
          description: "Identificador único (phone) do usuário recém-criado."
          example: "+558320548186"
...
  headers:
    Cookie:
      description: "Cookie HttpOnly contendo o token de refresh. O navegador o armazena automaticamente e o anexa às requisições subsequentes."
      schema:
        type: string
        example: "refreshToken=asdasda-asdasd-asdasd-asdasd..."
    ExpiredCookie:
      description: "Instrui o navegador a expirar e remover o cookie `refreshToken`. O valor do cookie é zerado e o `Max-Age` é definido como 0."
      schema:
        type: string
        example: "refreshToken=; Path=/api/auth; Max-Age=0; HttpOnly; Secure"
  responses:
    ErrorResponseInternalServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseDto"
          example:
            timestamp: "2025-08-15T13:55:00Z"
            status: 500
            message: "Erro interno do servidor. Por favor, tente novamente mais tarde."
            path: "/api/some/endpoint"
    AuthenticationExceptionResponse:
      description: "Exceção lançada quando o usuário não está autenticado ou o token é inválido/expirado."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseDto"
          example:
            timestamp: "2025-08-15T13:55:00Z"
            status: 401
            message: "Sua sessão é inválida ou expirou. Por favor, faça login novamente."
            path: "/api/auth/logout"
...
paths:
...
  /api/auth/login:
    post:
      operationId: "login"
      tags:
        - "Autenticação"
      summary: "Autenticar um usuário"
      description: "Autentica um usuário com nome de usuário e senha. Se as credenciais forem válidas, retorna um token de acesso (JWT) e um token de atualização (refresh token) em um cookie HttpOnly seguro."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDto"
      responses:
        "200":
          description: "Autenticação bem-sucedida. Retorna os dados do usuário e o token de acesso."
          headers:
            Cookie:
              $ref: "#/components/headers/Cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se os dados fornecidos não passarem na validação, como campos em branco ou com formato incorreto."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T14:21:00Z"
                status: 400
                message: "Erro de validação. Verifique os campos informados."
                path: "/api/auth/login"
                fieldErrors:
                  - fieldName: "username"
                    message: "O campo usuário deve ter entre 4 e 50 caracteres"
        "401":
          description: "Não autorizado. Retornado quando as credenciais (usuário/senha) são inválidas."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 401
                message: "Credenciais inválidas. Por favor, verifique seu usuário e senha."
                path: "/api/auth/login"
        "409":
          description: "Conflito de estado. Ocorre quando o estado atual da conta impede o login (ex: pendente de verificação, bloqueada ou desativada)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PendingVerification:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                    path: "/api/auth/login"
                Locked:
                  summary: "Conta bloqueada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                    path: "/api/auth/login"
                Disabled:
                  summary: "Conta desativada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está desativada."
                    path: "/api/auth/login"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/logout:
    post:
      operationId: "logout"
      tags:
        - "Autenticação"
      summary: "Realizar o logout do usuário"
      description: "Encerra a sessão do usuário. Invalida o refresh token no servidor e envia um cabeçalho para instruir o navegador a remover o cookie 'refreshToken'. Requer um token de acesso válido para ser executado."
      security:
        - bearerAuth: [ ]
      parameters:
        - in: "cookie"
          name: "refreshToken"
          required: false
          description: "O refresh token a ser invalidado. Normalmente, o navegador anexa este cookie automaticamente."
          schema:
            type: string
            example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
      responses:
        "204":
          description: "Logout realizado com sucesso. O refresh token foi invalidado e o cookie do navegador foi expirado."
          headers:
            ExpiredToken:
              $ref: "#/components/headers/ExpiredCookie"
        "400":
          description: "Requisição inválida. Ocorre se o formato do 'refreshToken' for inválido (não é um UUID válido)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 400
                message: "Formato inválido para o refresh token."
                path: "/api/auth/logout"
        "401":
          $ref: "#/components/responses/AuthenticationExceptionResponse"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/refresh-token:
    post:
      operationId: "refreshToken"
      tags:
        - "Autenticação"
      summary: "Renovar o token de acesso"
      description: "Gera um novo token de acesso (JWT) a partir de um refresh token válido. O refresh token deve ser enviado em um cookie HttpOnly chamado 'refreshToken'."
      parameters:
        - in: "cookie"
          name: "refreshToken"
          required: true
          description: "O refresh token recebido durante o login. O navegador anexa este cookie automaticamente."
          schema:
            type: string
            example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
      responses:
        "200":
          description: "Token de acesso renovado com sucesso. Um novo access token é retornado no corpo da resposta e um novo refresh token é enviado no cabeçalho 'Set-Cookie'."
          headers:
            Cookie:
              $ref: "#/components/headers/Cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se o formato do 'refreshToken' for inválido (não é um UUID válido)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T14:21:00Z"
                status: 400
                message: "Formato inválido para o refresh token."
                path: "/api/auth/refresh-token"
        "401":
          description: "Não autorizado. Ocorre se o 'refreshToken' estiver ausente, expirado ou não for encontrado no sistema."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingToken:
                  summary: "Refresh token ausente"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 401
                    message: "Sua sessão expirou. Por favor, faça login novamente."
                    path: "/api/auth/refresh-token"
                TokenNotFound:
                  summary: "Refresh token não encontrado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 401
                    message: "Refresh token não encontrado."
                    path: "/api/auth/refresh-token"
        "409":
          description: "Conflito de estado. Ocorre se a conta não estiver ativada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PendingVerification:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                    path: "/api/auth/refresh-token"
                Locked:
                  summary: "Conta bloqueada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                    path: "/api/auth/refresh-token"
                Disabled:
                  summary: "Conta desativada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está desativada."
                    path: "/api/auth/refresh-token"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/signup:
    post:
      operationId: "signup"
      tags:
        - "Autenticação"
      summary: "Cadastrar um novo usuário"
      description: "Registrar um novo usuário no sistema. A conta é criada com o status 'PENDING_VERIFICATION', e um código de verificação (OTP) é enviado para o telefone informado para posterior ativação."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequestDto"
      responses:
        "201":
          description: "Usuário cadastrado com sucesso e aguardando verificação."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se os dados de entrada forem inválidos (ex: senhas não coincidem, formato de telefone inválido ou violações de regras de campo)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PasswordsNotMatch:
                  summary: "Senhas não coincidem"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/signup"
                    fieldErrors:
                      - fieldName: "confirmPassword"
                        message: "A senha de confirmação não corresponde à senha."
                InvalidPhone:
                  summary: "Número de telefone inválido"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Número de telefone inválido. Verifique o DDD e a quantidade de dígitos."
                    path: "/api/auth/signup"
        "409":
          description: "Conflito de estado. Ocorre se o nome de usuário ou o telefone já estiverem cadastrados no sistema."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                UsernameExists:
                  summary: "Nome de usuário já em uso"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Esse nome de usuário já está em uso."
                    path: "/api/auth/signup"
                PhoneExists:
                  summary: "Número de telefone já em uso"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Esse número de telefone já está em uso."
                    path: "/api/auth/signup"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
...